Create Maximum XOR
You are given an array of
integers that denote the lucky
numbers of passengers that want
to board the flight. There are 4
planes available and all of them
are designed by Airbus. You need
to assign each passenger to one
of these four flights (it may be
possible that a ight remains
empty) such that the sum of the
happiness value of all the four
flights combined is maximum
possible. The happiness value of
a single flight is determined by
the xor value of the lucky
numbers of the passenger in it.
Input
The first line contains an integer
 as input denoting the size of
the array.
The next line contains space
separated integers denoting array
 .
Output
In the output, you need to print
the maximum flight happiness
possible.

example:
ip: 5
1 1 1 1 1
op=3